requirements.txt
Flask==0.10.1
speedtest-cli
hurry.filesize